Navbar update and router settings
Navbar has been updated. Added cart to navbar and redirected to cart page with Router. Footer was removed from lazy loading because it should appear on every page. Only page components were created for the necessary links in the navbar, but no codes were written. Eager loading was used on home screen and lazy loading was used on product screen.